fix: delete usage of contextBridge
Warning: I am aware that disabling contextIsolation and using simply window API is not considered secure. However this is a pattern that I have seen and used in other applications of kinect2 library, which is a (slightly deprecated) native module. I have not found other way to implement usage of Kinect API in Electron application, and this situation imposes special care of security in this project and also some restrictions (like usage of webpack).